Add a version number header to credential storage.
This adds a version number to credential storage in order to make
in-the-field upgrades easier.  If for some reason we need to extend
or change the `Credential` structure in a way that changes the CBOR,
we want to be able to continue to load the old format, and detect that
it needs updating.